Crime_against_women_in India 2001 - 2014
This data is collated from https://data.gov.in. It has state-wise and district level data on the various crimes committed against women between 2001 to 2014. Some crimes that are included are Rape, Kidnapping and Abduction, Dowry Deaths etc.

Here , Our main goal of this project is to classifying states as safe or unsafe. And also to understand whether the rate of Crime against Women has increased or decreased during the year 2001-2014 and which state or district has maximum crime rate.

Inferences and Conclusion India is one of the worst country for women in the world. And in recent time it is getting badder when crimes against women are happening in the name of their religion. According to scroll.in, 26 crimes against women are reported every hour, or one complaint every two minutes.
Andhra Pradesh had registered the most crimes against women from 2001 to 2014. Uttar Pradesh and West Bengal were second and third states with the most number of crimes against women. Andhra Pradesh also reported the highest number of cases of Insult to the modesty of Women.
Madhya Pradesh reported the most number of Rape cases and Assualt to her modest cases. Madhya Pradesh was also the fifth highest state with the most number of registered crime cases against women.
Uttar Pradesh reported the most number of Kidnapping and Abduction and Dowry Deaths. Seriously, still dowry!!!. The world is in the 21st century but UP is still in the 18th century.
Madhya Pradesh, Andhra Pradesh, Uttar Pradesh, Rajasthan and West Bengal account for almost half of all crimes committed against women around the country from 2001 to 2014.